refactor: create translations by phrase
using phrases in the code is easier to read, for instance if we translate a sentence, it would be nice to read that sentence in the jsx. however if we have such long phrases in our key value for createTranslations it becomes cumbersome to repeat the phrase multiple times. this commit will create translations by using the phrase as the key and repeat the language enums for each phrase as opposed to the other way around.